Display exercise statistics in Dashboard fragment (basic prtotype)

+++ app/src/main/java/com/example/fitnessfatality/data/models/routine/RoutineExercise.kt
@@ -21,7 +21,7 @@ data class RoutineExercise(
     val id: Long? = null,
     val routineId: Long,
     @ColumnInfo(name = "exercise_id")
-    val exerciseId: Int,
+    val exerciseId: Long,
     val sequenceOrder: Int = 0,
     val loggingParameters: HashMap<String, Int> = hashMapOf()
 ) {
